Moved regular components to separate files
