+++ frontend/src/common/views/fields/regular/TextViewRegular.jsx
@@ -1,12 +1,9 @@
 import React from 'react';
 import { Form } from 'react-bootstrap';
-import InlineView from './common/InlineView';
+import SingleInput from './common/SingleInput';
 
 const InputForm = ({
-  title,
-  text, setText,
-  isMultiline,
-  isBlocked,
+  text, setText, isMultiline, isBlocked,
   isInvalid,
 }) => {
   const dirty = text.length > 0;
@@ -15,11 +12,9 @@ const InputForm = ({
       <Form onSubmit={(e) => e.preventDefault()}>
         <Form.Control
           as={isMultiline ? 'textarea' : 'input'}
-          size="lg"
           rows={isMultiline ? 10 : 1}
           value={text}
-          placeholder={title}
-          onChange={(event) => { setText(event.target.value); }}
+          onChange={(event) => setText(event.target.value)}
           disabled={isBlocked}
           isInvalid={dirty && isInvalid}
         />
@@ -28,17 +23,17 @@ const InputForm = ({
   );
 };
 
-const TextViewInline = ({
-  title,
+const TextViewRegular = ({
+  title, description,
   isMultiline,
   input, setInput,
   highlighted,
   isBlocked,
   isInvalid,
 }) => (
-  <InlineView highlighted={highlighted}>
-    <InputForm title={title} text={input} setText={setInput} isMultiline={isMultiline} isBlocked={isBlocked} isInvalid={isInvalid} />
-  </InlineView>
+  <SingleInput title={title} description={description} highlighted={highlighted}>
+    <InputForm text={input} setText={setInput} isMultiline={isMultiline} isBlocked={isBlocked} isInvalid={isInvalid} />
+  </SingleInput>
 );
 
-export default TextViewInline;
+export default TextViewRegular;
